43 fix master

+++ src/main/java/com/progra3/proyecto/repository/VehiculoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.progra3.proyecto.entity.Vehiculo;
 
-public interface VehiculoRepository extends JpaRepository<Vehiculo,Long> {
+public interface VehiculoRepository extends JpaRepository<Vehiculo, Integer> {
 	
 }
